chore(platform): extract platform control center from chat-specific settings (#144)
Move platform-owned settings and control-center operations out of `chat.js` into dedicated platform API, model API, control, and notice modules.

Split platform and chat settings callbacks so the control center works without chat-specific hooks.

Route platform warnings and status feedback through the shell notice bar, including modal-safe dismissal and clearer fetch error reporting.

Closes #144
