+++ app/assets/chat.html
@@ -140,6 +140,8 @@
         </div>
       </header>
 
+      <div id="platformNotice" class="platform-notice" hidden aria-live="polite"></div>
+
       <section id="messages" class="messages"></section>
 
       <form id="composerForm" class="composer">
@@ -430,6 +432,10 @@
   <link rel="modulepreload" href="/assets/image-handler.js">
   <link rel="modulepreload" href="/assets/settings-ui.js">
   <link rel="modulepreload" href="/assets/chat-engine.js">
+  <link rel="modulepreload" href="/assets/platform-api.js">
+  <link rel="modulepreload" href="/assets/model-api.js">
+  <link rel="modulepreload" href="/assets/platform-notify.js">
+  <link rel="modulepreload" href="/assets/platform-controls.js">
     <script type="module" src="/assets/shell.js"></script>
 </body>
 </html>
